// src/lib/blockchain/types/SunflowerLandTrader.ts
export type Purchased = ContractEventLog<{
  sellerFarmId: string;
  buyerFarmId: string;
  listingId: string;
  resourceId: string;
  resourceAmount: string;
  sfl: string;
  0: string;
  1: string;
  2: string;
  3: string;
  4: string;
  5: string;
}>;